requirements.txt
discord.py>=2.3.0
impuestito>=0.4b0
python-dotenv>=1.0.0